# specs/openapi.yaml
openapi: 3.1.0
info:
  title: CS_tutoring BFF API
  version: v0.5.0
  description: 家教单发布与接单平台 BFF。唯一事实来源，与 specs/spec.md 第 3/6 部分同源。错误响应统一 {error, code, detail?}。v0.5.0 新增批量导入两端点（SPEC-002，specs/gig-import/spec.md §3）。
servers:
  - url: https://cstutoring.pages.dev
    description: 生产（Cloudflare Pages 同域）
  - url: http://127.0.0.1:8788
    description: 本地 wrangler pages dev
tags:
  - name: system
  - name: gigs
  - name: site-config
  - name: me

paths:
  /api/v1/healthz:
    get:
      tags: [system]
      summary: 健康检查
      operationId: getHealthz
      responses:
        '200':
          description: 正常
          content:
            application/json:
              schema:
                type: object
                required: [status, ts]
                properties:
                  status: { type: string, enum: [ok] }
                  ts: { type: string, format: date-time }
              example:
                status: ok
                ts: '2026-08-29T08:00:00.000Z'

  /api/v1/gigs:
    get:
      tags: [gigs]
      summary: 家教单列表（公开）
      operationId: listGigs
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [open, matched, closed, all], default: open }
          description: 缺省为 open；非法取值返回 422
        - name: grade_level
          in: query
          schema: { $ref: '#/components/schemas/GradeLevel' }
          description: 缺省不过滤
        - name: mode
          in: query
          schema: { $ref: '#/components/schemas/LessonMode' }
          description: 缺省不过滤
        - name: subject
          in: query
          schema: { type: string, maxLength: 40 }
          description: trim 后不区分大小写的精确匹配
        - name: q
          in: query
          schema: { type: string, maxLength: 60 }
          description: 标题（单号）搜索（v0.5.0）；trim 后空串视为未提供；对 title 做不区分大小写的包含匹配；与其他筛选参数 AND 叠加；含 * 或 % 返回 422（通配符不作为搜索语法）
        - name: district
          in: query
          schema: { $ref: '#/components/schemas/District' }
          description: 区县筛选（v0.4.0），缺省不过滤；非法取值返回 422
        - name: price
          in: query
          schema: { type: string, enum: [le50, '50-80', 80-120, 120-200, gt200] }
          description: 时薪档位筛选（v0.4.0），按 hourly_rate 过滤；hourly_rate 为 NULL 的单子不命中任何档位；非法取值返回 422
        - name: student_gender
          in: query
          schema: { type: string, enum: [male, female] }
          description: 学员性别筛选（v0.4.0）；unknown 表示未标注，不可作为筛选值；非法取值返回 422
        - name: sort
          in: query
          schema: { type: string, enum: [newest, rate_desc], default: newest }
          description: 排序（v0.4.0）；newest=created_at desc；rate_desc=hourly_rate 非空降序在前、NULL 殿后（殿后段保持 created_at desc）；非法取值返回 422
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: pageSize
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        '200':
          description: 列表；排序 created_at desc, id desc
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Gig' }
                  meta:
                    type: object
                    required: [page, pageSize, total]
                    properties:
                      page: { type: integer }
                      pageSize: { type: integer }
                      total: { type: integer }
              example:
                data:
                  - id: 550e8400-e29b-41d4-a716-446655440000
                    title: 高二数学一对一
                    subject: 数学
                    grade_level: senior
                    mode: online
                    region: 岳麓区·梅溪湖壹号
                    district: yuelu
                    hourly_rate: 150
                    student_gender: female
                    student_info: 女生，数学 85/150，基础较弱，计算易粗心
                    rate: 150/小时
                    schedule: 周六全天
                    requirements: 每周两次线上辅导，需要有耐心、能讲透基础题的学生
                    contact_wxid: null
                    status: open
                    published_by: 6f9619ff-8b86-d011-b42d-00c04fc964ff
                    created_at: '2026-08-29T08:00:00.000Z'
                    updated_at: '2026-08-29T08:00:00.000Z'
                meta: { page: 1, pageSize: 20, total: 1 }
        '422':
          $ref: '#/components/responses/ValidationError'
    post:
      tags: [gigs]
      summary: 发布家教单（admin）
      operationId: createGig
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GigCreate' }
            example:
              title: 高二数学一对一
              subject: 数学
              grade_level: senior
              mode: online
              region: 岳麓区·梅溪湖壹号
              district: yuelu
              hourly_rate: 150
              student_gender: female
              student_info: 女生，数学 85/150，基础较弱，计算易粗心
              rate: 150/小时
              schedule: 周六全天
              requirements: 每周两次线上辅导，需要有耐心、能讲透基础题的学生
      responses:
        '201':
          description: 已创建，status 为 open
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { $ref: '#/components/schemas/Gig' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationError' }

  /api/v1/gigs/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [gigs]
      summary: 家教单详情（公开，任意状态可读）
      operationId: getGig
      responses:
        '200':
          description: 详情；data 为 GigDetail（Gig + 发布者联系资料）
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { $ref: '#/components/schemas/GigDetail' }
        '404': { $ref: '#/components/responses/GigNotFound' }
    patch:
      tags: [gigs]
      summary: 修改家教单（admin；含状态流转）
      operationId: updateGig
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GigUpdate' }
            example: { status: matched }
      responses:
        '200':
          description: 已更新
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { $ref: '#/components/schemas/Gig' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/GigNotFound' }
        '422':
          description: 字段校验失败（VALIDATION_ERROR）或非法状态迁移（GIG_INVALID_TRANSITION）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              examples:
                validation:
                  value: { error: '请求参数不满足约束', code: VALIDATION_ERROR }
                invalidTransition:
                  value: { error: '非法的状态迁移', code: GIG_INVALID_TRANSITION }
    delete:
      tags: [gigs]
      summary: 删除家教单（admin）
      operationId: deleteGig
      security: [{ bearerAuth: [] }]
      responses:
        '204':
          description: 已删除，无响应体
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/GigNotFound' }

  /api/v1/gigs/import/preview:
    post:
      tags: [gigs]
      summary: 批量导入解析预览（admin；不写库）
      operationId: previewGigImport
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [raw_text]
              properties:
                raw_text: { type: string, minLength: 1 }
            example: { raw_text: '8.26长沙家教网10034639号家教\n学员地址：北部湾\n辅导科目：数学\n学员情况：初二、女' }
      responses:
        '200':
          description: 解析预览；行级 issues 标红、duplicate 标记（GigImportRow）
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: object
                    required: [rows]
                    properties:
                      rows:
                        type: array
                        items: { $ref: '#/components/schemas/GigImportRow' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationError' }

  /api/v1/gigs/import:
    post:
      tags: [gigs]
      summary: 批量写入家教单（admin；逐元素重校验，failed 不插入）
      operationId: commitGigImport
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [rows]
              properties:
                rows:
                  type: array
                  minItems: 1
                  items: { $ref: '#/components/schemas/GigCreate' }
      responses:
        '201':
          description: 合法行以 status=open、published_by=当前 admin 写入；failed 含批次内 index 便于前端定位
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GigImportCommit' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationError' }

  /api/v1/site-config:
    get:
      tags: [site-config]
      summary: 站点联系配置（公开）
      operationId: getSiteConfig
      responses:
        '200':
          description: 配置
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { $ref: '#/components/schemas/SiteConfig' }
              example:
                data:
                  wxid: admin-wx-001
                  qr_image_url: 'https://<project>.supabase.co/storage/v1/object/public/site-assets/qr.png'
                  notice: 报名请注明来意与想接的单子编号
    patch:
      tags: [site-config]
      summary: 修改站点联系配置（admin）
      operationId: updateSiteConfig
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SiteConfigUpdate' }
            example: { notice: 报名请注明来意与想接的单子编号 }
      responses:
        '200':
          description: 已更新
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { $ref: '#/components/schemas/SiteConfig' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationError' }

  /api/v1/me:
    get:
      tags: [me]
      summary: 查看当前登录账号资料（admin）
      operationId: getMe
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: 当前账号 profiles 行
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { $ref: '#/components/schemas/Profile' }
              example:
                data:
                  id: 6f9619ff-8b86-d011-b42d-00c04fc964ff
                  role: admin
                  display_name: null
                  avatar_url: null
                  wxid: my-wx-001
                  qr_image_url: null
                  created_at: '2026-08-29T08:00:00.000Z'
                  updated_at: '2026-08-29T08:00:00.000Z'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    patch:
      tags: [me]
      summary: 修改当前登录账号资料（admin）
      operationId: updateMe
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProfileUpdate' }
            example: { wxid: my-wx-001 }
      responses:
        '200':
          description: 已更新
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { $ref: '#/components/schemas/Profile' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationError' }

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase Auth JWT；BFF 服务端校验并查 profiles.role 判定 admin。
  schemas:
    GigStatus:
      type: string
      enum: [open, matched, closed]
    LessonMode:
      type: string
      enum: [online, offline]
    GradeLevel:
      type: string
      enum: [primary, junior, senior, college]
    District:
      type: string
      enum: [wangcheng, kaifu, yuelu, furong, tianxin, yuhua, changsha_county, other]
      description: 长沙区县枚举（v0.4.0）；other 兜底无明确区县归属（宁乡/浏阳等）。
    StudentGender:
      type: string
      enum: [male, female, unknown]
    Gig:
      type: object
      required: [id, title, subject, grade_level, mode, region, district, student_gender, student_info, requirements, status, published_by, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        title: { type: string, maxLength: 60 }
        subject: { type: string, maxLength: 40 }
        grade_level: { $ref: '#/components/schemas/GradeLevel' }
        mode: { $ref: '#/components/schemas/LessonMode' }
        region: { type: string, maxLength: 40 }
        district: { $ref: '#/components/schemas/District' }
        hourly_rate: { type: [integer, 'null'], minimum: 0, maximum: 10000 }
        student_gender: { $ref: '#/components/schemas/StudentGender' }
        student_info: { type: string, maxLength: 500 }
        rate: { type: [string, 'null'], maxLength: 40 }
        schedule: { type: [string, 'null'], maxLength: 120 }
        requirements: { type: string, maxLength: 2000 }
        contact_wxid: { type: [string, 'null'], maxLength: 40 }
        status: { $ref: '#/components/schemas/GigStatus' }
        published_by: { type: string, format: uuid }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    GigCreate:
      type: object
      required: [title, subject, grade_level, mode, region, district, student_info]
      properties:
        title: { type: string, minLength: 1, maxLength: 60 }
        subject: { type: string, minLength: 1, maxLength: 40 }
        grade_level: { $ref: '#/components/schemas/GradeLevel' }
        mode: { $ref: '#/components/schemas/LessonMode' }
        region: { type: string, minLength: 1, maxLength: 40 }
        district: { $ref: '#/components/schemas/District' }
        hourly_rate: { type: [integer, 'null'], minimum: 0, maximum: 10000 }
        student_gender: { $ref: '#/components/schemas/StudentGender' }
        student_info: { type: string, minLength: 1, maxLength: 500 }
        rate: { type: [string, 'null'], maxLength: 40 }
        schedule: { type: [string, 'null'], maxLength: 120 }
        requirements: { type: string, minLength: 1, maxLength: 2000 }
        contact_wxid: { type: [string, 'null'], maxLength: 40 }
      description: region、district 与 student_info 必填；hourly_rate 可空（面议/按次为 null）；student_gender 缺省 unknown（specs/spec.md §5.2）；未知字段忽略。
    GigUpdate:
      type: object
      minProperties: 1
      properties:
        title: { type: string, minLength: 1, maxLength: 60 }
        subject: { type: string, minLength: 1, maxLength: 40 }
        grade_level: { $ref: '#/components/schemas/GradeLevel' }
        mode: { $ref: '#/components/schemas/LessonMode' }
        region: { type: string, minLength: 1, maxLength: 40 }
        district: { $ref: '#/components/schemas/District' }
        hourly_rate: { type: [integer, 'null'], minimum: 0, maximum: 10000 }
        student_gender: { $ref: '#/components/schemas/StudentGender' }
        student_info: { type: string, minLength: 1, maxLength: 500 }
        rate: { type: [string, 'null'], maxLength: 40 }
        schedule: { type: [string, 'null'], maxLength: 120 }
        requirements: { type: string, minLength: 1, maxLength: 2000 }
        contact_wxid: { type: [string, 'null'], maxLength: 40 }
        status: { $ref: '#/components/schemas/GigStatus' }
      description: 对合并后的最终实体整体校验（region、district、student_info 合并后必须非空）；status 迁移受状态机约束。
    SiteConfig:
      type: object
      required: [wxid, qr_image_url]
      properties:
        wxid: { type: string, minLength: 1, maxLength: 40 }
        qr_image_url: { type: string, minLength: 1, maxLength: 500, pattern: '^https://' }
        notice: { type: [string, 'null'], maxLength: 200 }
    SiteConfigUpdate:
      type: object
      minProperties: 1
      properties:
        wxid: { type: string, minLength: 1, maxLength: 40 }
        qr_image_url: { type: string, minLength: 1, maxLength: 500, pattern: '^https://' }
        notice: { type: [string, 'null'], maxLength: 200 }
      description: notice 空字符串规范化为 null；未知字段忽略。
    GigImportDraft:
      type: object
      description: 导入草稿（解析/编辑中间结构，不入库）；未解析字段为 null 或空串（specs/gig-import/spec.md §3.1）。
      properties:
        title: { type: [string, 'null'], maxLength: 60 }
        subject: { type: [string, 'null'], maxLength: 40 }
        grade_level: { type: [string, 'null'], enum: [primary, junior, senior, college, null] }
        mode: { $ref: '#/components/schemas/LessonMode' }
        region: { type: [string, 'null'], maxLength: 40 }
        district: { $ref: '#/components/schemas/District' }
        hourly_rate: { type: [integer, 'null'], minimum: 0, maximum: 10000 }
        student_gender: { $ref: '#/components/schemas/StudentGender' }
        student_info: { type: [string, 'null'], maxLength: 500 }
        rate: { type: [string, 'null'], maxLength: 40 }
        schedule: { type: [string, 'null'], maxLength: 120 }
        requirements: { type: [string, 'null'], maxLength: 2000 }
        contact_wxid: { type: [string, 'null'], maxLength: 40 }
    FieldIssue:
      type: object
      description: 字段级问题；形状与错误响应 details 一致（specs/gig-import/spec.md §1）。
      required: [field, reason]
      properties:
        field: { type: string }
        reason: { type: string }
    GigImportRow:
      type: object
      required: [index, draft, issues, duplicate, status]
      properties:
        index: { type: integer, minimum: 0 }
        draft: { $ref: '#/components/schemas/GigImportDraft' }
        issues:
          type: array
          items: { $ref: '#/components/schemas/FieldIssue' }
        duplicate: { type: boolean }
        status: { type: string, enum: [ok, error] }
      description: status=error ⇔ issues 非空；duplicate=true 的行不计入可导入集合（specs/gig-import/spec.md §3.1）。
    GigImportCommit:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [created, failed]
          properties:
            created:
              type: array
              items: { $ref: '#/components/schemas/Gig' }
            failed:
              type: array
              items:
                type: object
                required: [index, code, details]
                properties:
                  index: { type: integer, minimum: 0 }
                  code: { type: string, enum: [VALIDATION_ERROR] }
                  details:
                    type: array
                    items: { $ref: '#/components/schemas/FieldIssue' }
    PublisherContact:
      type: object
      description: 发布者账号级联系资料（profiles 行）；学生联系弹层「发布者资料」回退层数据源。
      properties:
        wxid: { type: [string, 'null'], maxLength: 40 }
        qr_image_url:
          type: [string, 'null']
          maxLength: 500
          pattern: '^https://'
    GigDetail:
      allOf:
        - $ref: '#/components/schemas/Gig'
        - type: object
          required: [publisher_contact]
          properties:
            publisher_contact: { $ref: '#/components/schemas/PublisherContact' }
      description: GET /gigs/:id 专有形状；列表响应不含 publisher_contact。
    Profile:
      type: object
      required: [id, role, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        role: { $ref: '#/components/schemas/ProfileRole' }
        display_name: { type: [string, 'null'], maxLength: 40 }
        avatar_url: { type: [string, 'null'] }
        wxid: { type: [string, 'null'], maxLength: 40 }
        qr_image_url:
          type: [string, 'null']
          maxLength: 500
          pattern: '^https://'
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    ProfileRole:
      type: string
      enum: [admin, free]
    ProfileUpdate:
      type: object
      minProperties: 1
      properties:
        wxid: { type: [string, 'null'], maxLength: 40 }
        qr_image_url:
          type: [string, 'null']
          maxLength: 500
          pattern: '^https://'
      description: wxid 显式 null 合法（清空）；qr_image_url 只经 Storage 上传回写流程产生；未知字段忽略。
    Error:
      type: object
      required: [error, code]
      properties:
        error: { type: string }
        code: { type: string }
        detail: { type: string }
  responses:
    Unauthorized:
      description: 缺失或无效 JWT
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: 'Missing or invalid Authorization header', code: UNAUTHENTICATED }
    Forbidden:
      description: 已登录但非 admin
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: Forbidden, code: FORBIDDEN }
    GigNotFound:
      description: 单子不存在
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: 'Gig not found', code: GIG_NOT_FOUND }
    ValidationError:
      description: 参数校验失败
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Error'
              - type: object
                properties:
                  details:
                    type: array
                    items:
                      type: object
                      required: [field, reason]
                      properties:
                        field: { type: string }
                        reason: { type: string }
          example:
            error: '请求参数不满足约束'
            code: VALIDATION_ERROR
            details:
              - { field: region, reason: 'region 必填' }